raspi-configurator/requirements.txt
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
colorzero==1.1
cryptography==37.0.2
dbus-python==1.2.18
distro==1.5.0
gpiozero==1.6.2
idna==2.10
jeepney==0.8.0
pbkdf2==1.3
pycparser==2.21
PyGObject==3.38.0
python-apt==2.2.1
python-networkmanager==2.2
requests==2.25.1
RPi.GPIO==0.7.0
SecretStorage==3.3.2
six==1.16.0
smbus==1.1.post2
spidev==3.5
ssh-import-id==5.10
urllib3==1.26.5
